Move Icons to /lib
