+++ app/views/LegalView.js
@@ -13,7 +13,7 @@ import { isIOS, isAndroid } from '../utils/deviceInfo';
 import LoggedView from './View';
 import I18n from '../i18n';
 import { DARK_HEADER } from '../constants/headerOptions';
-import Icons from '../Icons';
+import Icons from '../lib/Icons';
 
 const styles = StyleSheet.create({
 	container: {
